+++ src/Server/Model/Scope.php
@@ -16,7 +16,7 @@
  * and is licensed under the MIT license.
  */
 
-namespace ZfrOAuth2\Server\Entity;
+namespace ZfrOAuth2\Server\Model;
 
 /**
  * A scope is associated to a token and define the permissions of the token
