+++ src/Server/ResourceServer.php
@@ -19,7 +19,7 @@
 namespace ZfrOAuth2\Server;
 
 use Psr\Http\Message\ServerRequestInterface;
-use ZfrOAuth2\Server\Entity\AccessToken;
+use ZfrOAuth2\Server\Model\AccessToken;
 use ZfrOAuth2\Server\Exception\InvalidAccessTokenException;
 use ZfrOAuth2\Server\Service\TokenService;
