+++ tests/Server/Container/AccessTokenServicefactoryTest.php
@@ -20,7 +20,7 @@ namespace ZfrOAuth2Test\Server\Container;
 
 use Interop\Container\ContainerInterface;
 use ZfrOAuth2\Server\Container\AccessTokenServiceFactory;
-use ZfrOAuth2\Server\Entity\AccessToken;
+use ZfrOAuth2\Server\Model\AccessToken;
 use ZfrOAuth2\Server\Options\ServerOptions;
 use ZfrOAuth2\Server\Repository\AccessTokenRepositoryInterface;
 use ZfrOAuth2\Server\Service\ScopeService;
