+++ tests/Server/Service/ClientServiceTest.php
@@ -18,7 +18,7 @@
 
 namespace ZfrOAuth2Test\Server\Service;
 
-use ZfrOAuth2\Server\Entity\Client;
+use ZfrOAuth2\Server\Model\Client;
 use ZfrOAuth2\Server\Repository\ClientRepositoryInterface;
 use ZfrOAuth2\Server\Service\ClientService;
